ref(shared): move CompiledFile read model into Shared
BuildFacadeInterface (in Shared) referenced Build\Domain\Compile\CompiledFile,
creating a Shared->Build dependency cycle. The DTO has no dependencies, so it
moves next to NamespaceInformation in Shared.
