+++ tests/php/Unit/Run/Application/DataReadersLoaderTest.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace PhelTest\Unit\Run\Application;
 
-use Phel\Build\Domain\Compile\CompiledFile;
 use Phel\Run\Application\DataReadersLoader;
+use Phel\Shared\CompiledFile;
 use Phel\Shared\Facade\BuildFacadeInterface;
 use Phel\Shared\NamespaceInformation;
 use PHPUnit\Framework\TestCase;
